fix: 엔티티 타입 불일치 수정

+++ module-api/src/main/java/com/whiskey/payment/controller/PaymentController.java
@@ -56,7 +56,7 @@ public class PaymentController {
         PaymentConfirmCommand command = new PaymentConfirmCommand(
             memberId,
             request.orderId(),
-            BigDecimal.valueOf(request.amount()),
+            request.amount(),
             request.paymentKey()
         );
